requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.6
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pyTelegramBotAPI==4.21.0
pytest==8.2.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.2.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
setuptools==70.2.0
